+++ src/components/qna/QaBadge/index.ts
@@ -1,0 +1,1 @@
+export * from './QaBadge'
