+++ src/components/qna/QuestionDetail/index.ts
@@ -1,0 +1,1 @@
+export * from './QuestionDetail'
